Move migration to it's own file
